Refactor to move launch button out of config questions component
